fix: third corrections

+++ src/controllers/rental.controller.js
@@ -165,6 +165,24 @@ export async function finalizeRental(req, res) {
 export async function deleteRental(req, res) {
   try {
     const { id } = req.params;
+
+    const checkRentalFinished = await db.query(
+      `
+      SELECT
+        *
+      FROM
+        rentals
+      WHERE
+        id = $1
+      AND
+        "returnDate" IS NOT NULL
+      ;
+    `,
+      [id]
+    );
+
+    if (checkRentalFinished.rowCount) return res.sendStatus(200);
+
     await db.query(
       `
       DELETE FROM
